Renamed u8timeseries to darts (#90)
* Renamed u8timeseries to darts

* correct name in RNN test

* update name in github workflows

* more renaming

* More replacements

* Update pip install instruction

* renamed conda recipe folder

* update pypi package name in README badges

* Temporarily name package u8darts

* Cosmetics

* Uniformize imports
