+++ .bumpversion.cfg
@@ -3,6 +3,6 @@ parse = (?P<major>\d+)\.(?P<minor>\d+)\.(?P<patch>\d+)|dev
 current_version = dev
 
 [bumpversion:file:setup.py]
-[bumpversion:file:u8timeseries/__init__.py]
-[bumpversion:file:conda_recipe/u8timeseries/meta.yaml]
+[bumpversion:file:darts/__init__.py]
+[bumpversion:file:conda_recipe/darts/meta.yaml]
 [bumpversion:file:docs/source/conf.py]
