+++ .github/workflows/develop.yml
@@ -1,4 +1,4 @@
-name: u8timeseries develop workflow
+name: darts develop workflow
 
 on:
   push:
@@ -45,7 +45,7 @@ jobs:
 
       - name: "6. Run linter"
         run: |
-          flake8 --config=setup.cfg u8timeseries
+          flake8 --config=setup.cfg darts
 
   docs:
     runs-on: ubuntu-latest
