+++ .gitignore
@@ -4,7 +4,7 @@ docs/build
 docs/source/examples
 docs/source/README.rst
 docs/source/generated_api
-u8timeseries.egg-info/
+darts.egg-info/
 build/
 dist/
 examples/.ipynb_checkpoints/
