+++ README.md
@@ -1,16 +1,16 @@
-# u8timeseries: Easy manipulation and forecasting of time series
+# darts: Easy manipulation and forecasting of time series
 
-![u8timeseries](https://github.com/unit8co/u8timeseries/raw/develop/static/images/logo-72dpi.png "u8timeseries") 
+![darts](https://github.com/unit8co/darts/raw/develop/static/images/logo-72dpi.png "darts") 
 
 ---
-[![PyPI version](https://badge.fury.io/py/u8timeseries.svg)](https://badge.fury.io/py/u8timeseries)
-![GitHub Workflow Status](https://img.shields.io/github/workflow/status/unit8co/u8timeseries/develop)
+[![PyPI version](https://badge.fury.io/py/u8darts.svg)](https://badge.fury.io/py/u8darts)
+![GitHub Workflow Status](https://img.shields.io/github/workflow/status/unit8co/darts/develop)
 ![Supported versions](https://img.shields.io/badge/python-3.6+-blue.svg)
-![Docker Image Version (latest by date)](https://img.shields.io/docker/v/unit8/u8timeseries?label=docker&sort=date)
-![PyPI - Downloads](https://img.shields.io/pypi/dm/u8timeseries)
-![GitHub Release Date](https://img.shields.io/github/release-date/unit8co/u8timeseries)
+![Docker Image Version (latest by date)](https://img.shields.io/docker/v/unit8/darts?label=docker&sort=date)
+![PyPI - Downloads](https://img.shields.io/pypi/dm/u8darts)
+![GitHub Release Date](https://img.shields.io/github/release-date/unit8co/darts)
 
-**u8timeseries** is a python library for easy manipulation and forecasting time series.
+**darts** is a python library for easy manipulation and forecasting time series.
 It contains a variety of models, from classics such as ARIMA to neural networks.
 The models can all be used in the same way, using `fit()` and `predict()` functions,
 similar to scikit-learn. The library also makes it easy to backtest models,
@@ -27,7 +27,7 @@ The following steps assume running inside a conda environment.
 If that's not possible, first follow the official instructions to install 
 [fbprophet](https://facebook.github.io/prophet/docs/installation.html#python)
 and [torch](https://pytorch.org/get-started/locally/), then skip to 
-[Install u8timeseries](#install-u8timeseries)
+[Install darts](#install-darts)
 
 To create a conda environment for Python 3.7
 (after installing [conda](https://docs.conda.io/en/latest/miniconda.html)):
@@ -47,9 +47,9 @@ Don't forget to activate your virtual environment
 
     conda install -c conda-forge -c pytorch pip fbprophet pytorch cpuonly
 
-### Install u8timeseries
+### Install darts
 
-    pip install u8timeseries
+    pip install u8darts
 
 ### Running the examples only, without installing:
 
@@ -70,7 +70,7 @@ For this setup to work you need to have a Docker service installed. You can get 
 Create `TimeSeries` object from a Pandas DataFrame, and split in train/validation series:
 
 ```python
-from u8timeseries import TimeSeries
+from darts import TimeSeries
 
 df = pd.read_csv('AirPassengers.csv', delimiter=",")
 series = TimeSeries.from_dataframe(df, 'Month', '#Passengers')
@@ -80,7 +80,7 @@ train, val = series.split_after(pd.Timestamp('19590101'))
 Fit an exponential smoothing model, and make a prediction over the validation series' duration:
 
 ```python
-from u8timeseries import ExponentialSmoothing
+from darts import ExponentialSmoothing
 
 model = ExponentialSmoothing()
 model.fit(train)
@@ -95,13 +95,13 @@ plt.legend()
 plt.xlabel('Year')
 ```
 
-![example](https://github.com/unit8co/u8timeseries/raw/develop/static/images/example.png "example") { width=100% }
+![example](https://github.com/unit8co/darts/raw/develop/static/images/example.png "example") { width=100% }
 
 We invite you to go over the example notebooks in the `examples` directory.
 
 ## Documentation
 
-The documentation of the API and models is available [here](https://unit8co.github.io/u8timeseries/).
+The documentation of the API and models is available [here](https://unit8co.github.io/darts/).
 
 ## Features
