+++ conda_recipe/environment.yml
@@ -1,5 +1,5 @@
 # conda-specific dependencies for the dev environment
-name: u8timeseries-dev
+name: darts-dev
 dependencies:
   - python>=3.6
   - conda-build
