+++ darts/models/rnn_model.py
@@ -124,7 +124,7 @@ class RNNModel(TorchForecastingModel):
         model
             Either a string specifying the RNN module type ("RNN", "LSTM" or "GRU"),
             or a PyTorch module with the same specifications as
-            `u8timeseries.models.rnn_model.RNNModule`.
+            `darts.models.rnn_model.RNNModule`.
         output_length
             Number of time steps to be output by the forecasting module.
         hidden_size
