+++ darts/tests/test_metrics.py
@@ -4,7 +4,7 @@ import pandas as pd
 import logging
 
 from ..timeseries import TimeSeries
-from u8timeseries.metrics import metrics
+from ..metrics import metrics
 
 
 class MetricsTestCase(unittest.TestCase):
