+++ darts/tests/test_timeseries_generation.py
@@ -2,7 +2,7 @@ import unittest
 
 import pandas as pd
 
-from u8timeseries.utils.timeseries_generation import (
+from ..utils.timeseries_generation import (
     constant_timeseries,
     linear_timeseries,
     sine_timeseries,
