+++ scripts/build_docker.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
-docker build -t unit8/u8timeseries:latest ../
+docker build -t unit8/darts:latest ../
